requirements.txt
boto3==1.40.40
botocore==1.40.40
jmespath==1.0.1
mysql-connector-python==9.4.0
python-dateutil==2.9.0.post0
s3transfer==0.14.0
six==1.17.0
urllib3==2.5.0
